Refactor and clean up constants
First, we make constants actually constants, instead of global variables.
Second, TotalBalance and TotalBalanceLength were not used, and there is little reason to expose them
if we give that responsability to the Amount type.
Last, the genesis command was using TotalBalanceWithComma as initial value, but it was its only use.
To make it clearer and decouple usages, we move and rename this constant to the genesis module