Use log15 builtin StdoutHandler
The isatty check didn't work when `docker run` was used,
even when the `--tty` option was used.
As a result, most of the debug wasn't visible
when running within a docker container.

log15 already provide an handler that just works to do exactly this,
so we depend on this instead.