Add basic integration tests
This adds the basic structure for integration tests,
as well as an integration test for account creation.

As opposed to `go test`, this is an end-to-end test which uses the production binary.
It does not rely on sebak in any way for testing, instead we store hardcoded requests
that we send to sebak, and observe the result.

It will need a few improvements to make it scalable:
- There are random sleeps scattered in the code base
  We should probably use a more sensible retry mechanism instead
- Generating requests is time consuming.
  We would greatly benefit from a way to log all traffic which is not Wireshark.
- Sending parallel requests is not yet supported
- This might grow into a monster of a shell script