Fix #190: `exec` sebak to properly handle signals
When Docker starts the container, the main process is the `sh` process running `entrypoint.sh`.
As a result, when one runs `docker stop ${CONTAINER_NAME}`, the signal is sent to `sh`,
which doesn't forward it to `sebak`, leading to us not being able to gracefully stop and `SIGKILL`
being sent after the timeout (10s by default).
Using `exec` we replace the `sh` process with `sebak`,
solving this issue in the simplest way.