LevelDB: Properly close the storage on DB close
When using the file storage, GoLevelDB acquires a FileLock to avoid concurrent open / data corruption.
The finalizer of the instance is set to release that file lock.
However, we do not store an instance of this storage, and the DB instance doesn't know it should Close it
when its own Close method is called.
Fortunately, DB provides the OpenFile method which informs the DB instance it should close the file on DB.Close.