Do not reconnect every second after the initial connection
Once we are connected to a node, stop sending `/connect` messages.
By setting an `idleTimeout` of `0` and `keepAlive` to `true`
we ensure that the connection is never closed.
We're still suffering from bugs when a node comes back up,
but since Sebak does not currently support it,
it isn't a problem.