Remove Node.SetKeyPair function
It doesn't make sense to have it, as not all nodes have this functionality.
When it is needed, we can just provide it in the constructor.
This is made obvious by the diff, where there was a pattern of constructing
the object with `kp.Address()` then calling the just-constructed object with
`kp` as argument to `SetKeyPair`.