Remove ability to UnmarshalJSON a LocalNode
It does not make sense to UnmarshalJSON a LocalNode,
as we should never need to turn a JSON string into a LocalNode instance.
A LocalNode instance owns a keypair.Full which we should never serialize and send for obvious security reason.
On the other hand, it does not make sense to have a LocalNode without keypair.Full.
However, we leave the ability to serialize a LocalNode to a Validator.