Replace checkpoints with sequence IDs
As we are transitioning from checkpoints to sequence ID,
one of the main change is that account will start with a sequence ID of 0.
In order to break down that change, account now start with a checkpoint of `0`s.

The second part of the change is to remove next checkpoints.
Next checkpoint do not play a role in the consensus. They take up space and are solely debugging information.
Their original role was to provide a stronger rule w.r.t balance verification,
however such rules proved to be impossible to implement without compromising
other aspects of the design (namely, it made issuing transactions very complicated).
Also removes the ability to get Block{Operation,Transaction} by checkpoint,
as it's not used anyway and doesn't make sense with sequence id.

Lastly everything was renamed, including some names being shortened.