Use a list of all validators to calculate the proposer
To calculate the proposer, we should obviously not depend on the node state,
but something that is shared by everyone.
At the moment we can work under the assumption that the validator list will be the same for everyone.
However, there's no way we can assume that the list of connected nodes is the same for everyone,
or that we should base any part of the consensus on that.