Do not perform checks inside of other goroutines & fix data race
Unfortunately, require.XXX cannot be used in a goroutine started by a test goroutine,
and has to be used in the test one.
This means that any code that consume events and check the result must not run in another goroutine.

This change also fixes a data race, where some data was accessed concurrently from 2 goroutines.
An array was appended to, and read from another goroutine, however synchronization was not
properly done and the race detector was complaining.