Add conf for ISAAC timeout features (#292)
### Background
This is second part of the new ISAAC implementation.
According to the new ISAAC design, 4 Timeout features are needed.
And I guess, the number of transactions in one block is needed to be set because we don't know the appropriate number yet. So we should experiment with these 5 features(4 timeout and 1 n of tx).

### Solution
Add these features.

### Comment
NodeRunnerConfiguration is included in NodeRunnerStateManager and it is set in run.go:331. NodeRunnerStateManager will be uploaded within next pr.