MemoryNetwork: Remove global state
This global state was dangerous as test cases can be run in parallel.
It was actually an issue, as the comment in `CleanUpMemoryNetwork` shows.
By removing the global state we ensure that this kind of error is less likely
to happen: although races are still possible within a test cases,
they are now impossible accross test cases.