checker's finishBallot: Don't use named return
When returning an error from a function, it's also better to provide an unusable value to the user.
This way, if error checking is ignored, it is more likely that the erronerous value triggers
more errors down the road, leading to the issue being noticed nonetheless.
This is another kind of defensive programming.