Remove the need to pass the testing object to helper functions
Those test helpers function are generating data to make testing easier.
In principle, they should never fail, and if they do, there's a big bad bug somewhere.
We should not be using the testing object for those kinds of checks, as they are not part of the tests.
Moreover, if that bbb (big bad bug!) happens in another goroutine, it will be ignored
with the testing object, bug cause a crash with panic.