Integration Testing Env for SDK. (#486)
Github Issue
#471

To prove features working, it is very important to make a test scenario in a client position.
Existing integration test which is written in shell script with curl, jq, ... etc is not enough to achieve the goal
Solution
Build a client module. By using the client module, make a test scenario in client-side. (The scenario is not written yet)

The client module should be separated in another project. But for now, it needed to be built in same repository. Hence, the client-side container share same build output from sebak. That is why the Dockerfile(s) are revised.