Consensus rules cannot be broken in finishOperation
The consensus rules are checked in checker_ballot_transaction's `ValidateOp`.
So if this error happens, it means there's an important bug in sebak,
and that the consensus rules were bypassed.
As a result, we remove the checks which are not necessary,
and panic on operations that shouldn't return an error.