Improve TestBlockHeightOrdering
This test checks for an impossible scenario: blocks being saved out of order.
It's not something that should ever happen, and if anything, this test should not pass.

During the review, it was pointed out that the test is meants to test gets
out of order, not puts out of order, so the test was changed to do just that.