node/runner/api: Panic instead of checking error for test utilities
For test utility, it is better to straight up panic than do an error checking.
Since those errors should never happen, and the overwhelming majority of the case
happens after a change to the code, it is better to have the error as close
to the place where it is detected as possible, which panic achieves.