Move all dependency to keypair within Sebak
Keypair is at the core of Sebak, but currently the dependency is external,
and scattered around the packages.
By putting it in a central place, we'll reduce the impact of any upstream change.
Additionally, we can provide convenience functions (e.g. the non-erroring Random)
that are better suited to Sebak's needs.