Make the node save block in ALLCONFIRM
Fixes #637
If there are too many transactions in a ballot,
ballot is expired after saving block.
Because state transition is called after saving transactions.
Confirming process is,
1. ACCEPT ballot exceeded
1. save block
1. save transaction
1. state transit to ALLCONFIRM
1. state transit to INIT

If there are many transactions, a timeout might occur while
saving them(2).

Therefore, I've modified it to this process
1. ACCEPT ballot exceeded
1. state transit to ALLCONFIRM
1. save block
1. save transaction
1. state transit to INIT

And there is 1 minute timeout in ALLCONFIRM. It's because of exceptional
case.