Remove support for Serializable
The use of the Serializable interface means there is only one way to serialize a data structure.
However, in different situations, we might want to use different strategies,
such as JSON serialization for network communication,
and binary for on-disk storage.

As a result, instead of having Serializable.Serialize call json.Marshal,
we move the json.Marshal calls to the network part.
This is also more consistent with Go's standard library marshalling interfaces.