Use a byte for OperationType
Instead of storing a string in memory, which takes up quite some space,
we store a single byte and provide a mean to convert to/from string.

We still store a string on disk because we need to provide a seamless
way to convert blocks, as we need to support existing deployments.