Remove Serialize method from Operation interface
Serialize is not useful, as it is solely used to serialize to JSON.
It creates problem when one wants to serialize in different formats,
e.g. JSON for network communication, and binary for on-disk storage.
Removing this interface allows us to move the choice of the serialization
format to the caller of the method, where it actually makes more sense.