Operation: Use slightly more clever code to deserialize body
Instead of copy-pasting blocks of code inside a `switch`,
we can just have a function instantiating the type and
greatly reduce duplication.

As `Body` is defined to be an interface to value instead
of an interface to pointer, a bit of magic is needed
with the `reflect` module to make it work.