Enable binary serialization of BlockOperation (+ add upgrade command)
In the LevelDB serialization code, `BinaryMarshaller` takes priority over `TextMarshaller`.
This allows us to simply switch on binary serialization with a few lines of code.

However, it is a breaking change for existing deployments,
as old text-serialized data won't be deserializable.

There are multiple way to approach this problem, among other things:
- Trying one of the method, then falling back if it doesn't work
- Setting a given block height in the future for the transition
- Replay all the transaction data to recreate the DB
- Force conversion of all the existing data

We choose, for simplicity, and to help reducing existing load on nodes,
to go with the later approach.
We introduce `sebak upgrade`, which will do the job only if needed.
In addition, a check was introduced in `run`,
to avoid the node starting with the new code and old data,
which could lead to memory corruption or crash.

Note that the code in `sebak upgrade` is not guaranteed to work in a future release.
Doing so would require specifying exactly what transformation is to be made,
which would overly complicate the process for little benefit.
It can be done in the future if the user base justifies the need.