client: Improve StreamTransaction and merge in StreamTransactionByHash
Adding the `ids` as variadic array of string allows to listen to more than one transactions,
which was not possible with `StreamTransactionsByHash`.
It also allows us to merge the two functions for a simpler/leaner interface.
Additionally, the return value for `json.Marshal` was never checked,
as it was overriden in the return statement.
Last but not least, add documentation for users.