Send initial values when accounts are /subscribed
A common use case is to watch an account waiting for a specific update.
So the client would fire a request, and then /subscribe
to the account's key, waiting for the operation to be included
in the chain.
However, that would not always work, e.g. if the request
is completed before the `/subscribe` is received,
and no other action is taken on the account,
the client would wait forever.
As a band-aid, we send the initial value of the account(s)
when accounts are watched, if the is any value.