Replace NewEndpointFromString with ParseEndpoint, improve tests
NewEndpointFromString was primarily used from tests and had a different semantic
than the more complere ParseEndpoint.
Since having two functions to convert a string to an Endpoint is a bad sign,
the lesser function (with the longest name) was removed.
In addition, since many (mosts?) tests seemed to ignore the error return value,
a test-only MustParseEndpoint was introduced, as ignoring a return value
is a deadly sin.